chore: Turn transpilation on in the test environment
As I am using Webpack, which supports ES6 modules, I configured Babel to not transpile them.
That's because if Babel transpiled ES6 modules, we would lose a Webpack feature called `tree shaking`, that is an optimization made by Webpack when it bundles our modules.
But with transpilation of ES6 modules turned off with the option {"modules": false} - added in the previous commit -, I had to turn it on again in the test environment.