feat: Add password encryption and sign user id
Added password encryption to User model.
Changed the JSON web token payload to id, instead of username.
Modified the `query` unit testing module to target user instead of instructor, to test the new methods added to the User model (it will also become handy in the future, as Instructor will probably be dropped).